allow CMD-SHELL healthchecks
