+++ src/pkg/cli/client/byoc/baseclient.go
@@ -68,12 +68,17 @@ func NewByocBaseClient(ctx context.Context, grpcClient client.GrpcClient, tenant
 		PulumiStack:   "beta", // TODO: make customizable
 		Quota: quota.Quotas{
 			// These serve mostly to pevent fat-finger errors in the CLI or Compose files
-			Cpus:       16,
-			Gpus:       8,
-			MemoryMiB:  65536,
-			Replicas:   16,
-			Services:   40,
-			ShmSizeMiB: 30720,
+			ServiceQuotas: quota.ServiceQuotas{
+				Cpus:       16,
+				Gpus:       8,
+				MemoryMiB:  65536,
+				Replicas:   16,
+				ShmSizeMiB: 30720,
+			},
+			ConfigCount: 20,   // TODO: add validation for this
+			ConfigSize:  4096, // TODO: add validation for this
+			Ingress:     10,   // TODO: add validation for this
+			Services:    40,
 		},
 		bootstrapLister: bl,
 	}
